Created splash screen image subject to be changed. Moved all the home activities to a separate class.
