+++ app/src/main/java/uk/ac/tees/cupcake/Home/HomeFragment.java
@@ -1,20 +1,20 @@
-package uk.ac.tees.cupcake;
+package uk.ac.tees.cupcake.Home;
 
+import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-public class NewsFeedFragment extends Fragment {
+import uk.ac.tees.cupcake.R;
 
+public class HomeFragment extends Fragment {
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
 
-        return  inflater.inflate(R.layout.newsfeed_fragment,container,false);
+        return  inflater.inflate(R.layout.home_fragment,container,false);
     }
 }
