+++ app/src/main/res/drawable/background.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item android:drawable="@color/grey" />
+
+    <item>
+        <bitmap android:src="@drawable/cupcake"
+            android:gravity="center"/>
+    </item>
+</layer-list>
