+++ app/src/main/res/layout/home_center_viewpager.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".HomeActivity"
+    tools:context=".Home.HomeActivity"
     tools:showIn="@layout/app_bar_main">
 
         <android.support.v4.view.ViewPager
